requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pdfminer.six==20221105
python-docx==0.8.11
pydantic==1.10.12
requests==2.32.0
